Add admin to do for trust profiles
AI's will need to maintain behaviour histories for users and/of their anonymous identified, so as to be able to recommend who can or can not be trusted, and whether their current or future behavior is likely to be incongruent or toxic.  Such profiles would measure many properties of trustability, and may need to be irrevocable/undeletable, even if abandoned, like a rap sheet for good and bad, but such sheets are typically considered inapplicable to minors, yet any anonymous user could be a minor.